backend/requirements.txt
Flask
pycryptodome
jsonschema
pymongo
requests
PyJWT
gunicorn
pytest
Flask-Testing
requests-mock
cryptography

# Documentation
sphinx
sphinx-markdown-builder